LayoutTests/imported/w3c/web-platform-tests/cors/304-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/304.py?id=1&r=%s1622137963472
Blocked access to external URL http://www1.localhost:8800/cors/resources/304.py?id=3&r=%s1622137963474
Blocked access to external URL http://www1.localhost:8800/cors/resources/304.py?id=5&r=%s1622137963474
Blocked access to external URL http://www1.localhost:8800/cors/resources/304.py?id=7&r=%s1622137963474
CORS - 304 Responses


FAIL A 304 response with no CORS headers inherits from the stored response assert_equals: didn't get successful 1st response; expected "Success" but got ""
FAIL A 304 can expand Access-Control-Expose-Headers assert_equals: didn't get successful 1st response; expected "Success" but got ""
FAIL A 304 can contract Access-Control-Expose-Headers assert_equals: didn't get successful 1st response; expected "Success" but got ""
FAIL A 304 can change Access-Control-Allow-Origin assert_equals: didn't get successful 1st response; expected "Success" but got ""
